refactor(interfaces): base class, configurable lidar and TQC images
Add TrackMania2020InterfaceBase with OpenPlanet client hook and shared
control/reset/image-buffer logic. TM2020Interface delegates to the base
and drops duplicate embedded lidar subclasses.

TM2020InterfaceLidarConfigurable unifies lidar, lidar+progress, and
lidar+progress+camera via include_progress and include_camera_images;
thin subclasses preserve existing public class names.

TM2020InterfaceTQC gains include_camera_images; TM2020InterfaceTQCWithImages
delegates to it. TQC uses _build_openplanet_client instead of pre-assigning
client before initialize_common.

Made-with: Cursor
